refactor(wallet-mobile): moving things around
