+++ apps/wallet-mobile/.storybook/storybook.requires.js
@@ -271,8 +271,8 @@ const getStories = () => {
     "./src/HW/LedgerTransportSwitchModal/LedgerTransportSwitchModal.stories.tsx": require("../src/HW/LedgerTransportSwitchModal/LedgerTransportSwitchModal.stories.tsx"),
     "./src/kernel/metrics/metrics.stories.tsx": require("../src/kernel/metrics/metrics.stories.tsx"),
     "./src/legacy/Modal/Modal.stories.tsx": require("../src/legacy/Modal/Modal.stories.tsx"),
-    "./src/NftDetails/NftDetails.stories.tsx": require("../src/NftDetails/NftDetails.stories.tsx"),
-    "./src/NftDetails/NftDetailsImage.stories.tsx": require("../src/NftDetails/NftDetailsImage.stories.tsx"),
+    "./src/Nfts/NftDetails/NftDetails.stories.tsx": require("../src/Nfts/NftDetails/NftDetails.stories.tsx"),
+    "./src/Nfts/NftDetails/NftDetailsImage.stories.tsx": require("../src/Nfts/NftDetails/NftDetailsImage.stories.tsx"),
     "./src/Nfts/Nfts.stories.tsx": require("../src/Nfts/Nfts.stories.tsx"),
     "./src/Nfts/NoNftsScreen.stories.tsx": require("../src/Nfts/NoNftsScreen.stories.tsx"),
     "./src/Staking/DelegationConfirmation/DelegationConfirmation.stories.tsx": require("../src/Staking/DelegationConfirmation/DelegationConfirmation.stories.tsx"),
