+++ apps/wallet-mobile/src/Nfts/NftDetails/NftDetails.stories.tsx
@@ -1,10 +1,10 @@
 import {storiesOf} from '@storybook/react-native'
 import React from 'react'
 
-import {QueryProvider, RouteProvider} from '../../.storybook/decorators'
-import {MediaDetails} from '../features/Portfolio/common/MediaDetails/MediaDetails'
-import {SelectedWalletProvider} from '../features/WalletManager/context/SelectedWalletContext'
-import {mocks, nft} from '../yoroi-wallets/mocks'
+import {QueryProvider, RouteProvider} from '../../../.storybook/decorators'
+import {MediaDetails} from '../../features/Portfolio/common/MediaDetails/MediaDetails'
+import {SelectedWalletProvider} from '../../features/WalletManager/context/SelectedWalletContext'
+import {mocks, nft} from '../../yoroi-wallets/mocks'
 
 storiesOf('NFT/Details', module)
   .add('Default', () => {
