+++ apps/wallet-mobile/src/features/Auth/CreatePinInput/CreatePinInput.stories.tsx
@@ -3,7 +3,7 @@ import {storiesOf} from '@storybook/react-native'
 import {AsyncStorageProvider} from '@yoroi/common'
 import React from 'react'
 
-import {rootStorage} from '../../../yoroi-wallets/storage/rootStorage'
+import {rootStorage} from '../../../kernel/storage/rootStorage'
 import {CreatePinInput} from './CreatePinInput'
 
 storiesOf('CreatePinInput', module).add('Default', () => (
