+++ apps/wallet-mobile/src/features/Dev/DeveloperScreen.tsx
@@ -15,11 +15,11 @@ import {agreementDate} from '../../kernel/config'
 import {showErrorDialog} from '../../kernel/dialogs'
 import {errorMessages} from '../../kernel/i18n/global-messages'
 import {AppRoutes, useWalletNavigation} from '../../kernel/navigation'
-import {storageVersionMaker} from '../../migrations/storageVersion'
+import {storageVersionMaker} from '../../kernel/storage/migrations/storageVersion'
+import {rootStorage} from '../../kernel/storage/rootStorage'
 import {isEmptyString} from '../../utils/utils'
 import {generateAdaMnemonic} from '../../yoroi-wallets/cardano/mnemonic'
 import {useCreateWallet} from '../../yoroi-wallets/hooks'
-import {rootStorage} from '../../yoroi-wallets/storage/rootStorage'
 import {NetworkId} from '../../yoroi-wallets/types'
 import {useAuth} from '../Auth/AuthProvider'
 import {useLegalAgreement, useResetLegalAgreement} from '../Initialization/common'
