+++ apps/wallet-mobile/src/features/Receive/common/SingleOrMultipleAddressesModal/SingleOrMultipleAddressesModal.stories.tsx
@@ -3,8 +3,8 @@ import {storiesOf} from '@storybook/react-native'
 import {AsyncStorageProvider} from '@yoroi/common'
 import * as React from 'react'
 
+import {rootStorage} from '../../../../kernel/storage/rootStorage'
 import {mocks} from '../../../../yoroi-wallets/mocks'
-import {rootStorage} from '../../../../yoroi-wallets/storage/rootStorage'
 import {SelectedWalletProvider} from '../../../WalletManager/context/SelectedWalletContext'
 import {SingleOrMultipleAddressesModal} from './SingleOrMultipleAddressesModal'
