+++ apps/wallet-mobile/src/kernel/storage/rootStorage.ts
