+++ apps/wallet-mobile/src/yoroi-wallets/cardano/byron/ByronWallet.ts
@@ -19,9 +19,10 @@ import {WalletMeta} from '../../../features/WalletManager/common/types'
 import walletManager from '../../../features/WalletManager/common/walletManager'
 import LocalizableError from '../../../kernel/i18n/LocalizableError'
 import {logger} from '../../../kernel/logger/logger'
+import {makeWalletEncryptedStorage, WalletEncryptedStorage} from '../../../kernel/storage/EncryptedStorage'
+import {Keychain} from '../../../kernel/storage/Keychain'
 import {HWDeviceInfo} from '../../hw'
 import {makeMemosManager, MemosManager} from '../../memos'
-import {Keychain, makeWalletEncryptedStorage, WalletEncryptedStorage} from '../../storage'
 import {
   AccountStateResponse,
   BackendConfig,
