+++ apps/wallet-mobile/src/yoroi-wallets/cardano/shelley/tests/testnet/ShelleyWalletTestnet.test.ts
@@ -4,9 +4,9 @@ import AsyncStorage from '@react-native-async-storage/async-storage'
 import {Balance} from '@yoroi/types'
 
 import {WalletMeta} from '../../../../../features/WalletManager/common/types'
+import {EncryptedStorage, EncryptedStorageKeys} from '../../../../../kernel/storage/EncryptedStorage'
+import {rootStorage} from '../../../../../kernel/storage/rootStorage'
 import {HWDeviceInfo} from '../../../../hw'
-import {EncryptedStorage, EncryptedStorageKeys} from '../../../../storage'
-import {rootStorage} from '../../../../storage/rootStorage'
 import {DefaultAsset} from '../../../../types'
 import {ShelleyAddressGeneratorJSON} from '../../../chain'
 import {WalletJSON} from '../../../shelley/ShelleyWallet'
