+++ apps/wallet-mobile/src/yoroi-wallets/cardano/transactionManager/transactionManager.test.ts
@@ -2,7 +2,7 @@ import AsyncStorage from '@react-native-async-storage/async-storage'
 import {fromPairs} from 'lodash'
 import DeviceInfo from 'react-native-device-info'
 
-import {rootStorage} from '../../storage/rootStorage'
+import {rootStorage} from '../../../kernel/storage/rootStorage'
 import {Transaction} from '../../types/other'
 import {ApiHistoryError} from '../errors'
 import {
