+++ apps/wallet-mobile/src/yoroi-wallets/memos/memosManager.test.ts
@@ -1,6 +1,6 @@
 import AsyncStorage from '@react-native-async-storage/async-storage'
 
-import {rootStorage} from '../storage/rootStorage'
+import {rootStorage} from '../../kernel/storage/rootStorage'
 import {makeMemosManager} from './memosManager'
 
 describe('memos manager', () => {
