--- apps/wallet-mobile/src/yoroi-wallets/storage/index.ts
@@ -1,3 +1,0 @@
-export * from './EncryptedStorage'
-export * from './Keychain'
-export * from './KeychainStorage'
