+++ apps/wallet-mobile/translations/messages/src/Nfts/NftsNavigator.json
@@ -6,12 +6,12 @@
     "start": {
       "line": 47,
       "column": 9,
-      "index": 1390
+      "index": 1389
     },
     "end": {
       "line": 50,
       "column": 3,
-      "index": 1461
+      "index": 1460
     }
   }
 ]
